+++ index.html
@@ -230,5 +230,5 @@
   </template>
 </body>
 
-<script src="/js/main.js"></script>
+<script src="/js/main.js" type="module"></script>
 </html>
